ZEPPELIN-1430. Display appId and webui link in LivyInterpreter's output
### What is this PR for?
For now, it is hard to figure out what the yarn application of the livy session represent, it would be better to display the appId and webui link in the output of LivyInterpreter for diagnosing purpose. It can also be applied to the native SparkInterpreter, but could be done in another ticket.

### What type of PR is it?
[Feature]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-1430

### How should this be tested?
Tested manually

### Screenshots (if appropriate)
![image](https://cloud.githubusercontent.com/assets/164491/18463333/e4eab580-79bb-11e6-8c8d-393ab6805638.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #1426 from zjffdu/ZEPPELIN-1430 and squashes the following commits:

88009cb [Jeff Zhang] update doc
459f75e [Jeff Zhang] fix typo
eb7ec27 [Jeff Zhang] add flag to enable display app info in frontend
a087a1d [Jeff Zhang] ZEPPELIN-1430. Display appId and webui link in LivyInterpreter's output