ZEPPELIN-1562 updated api doc
What is this PR for?

The URL for running a paragraph synchronously using REST api is mistakenly given as "http://[zeppelin-server]:[zeppelin-port]/api/notebook/job/[notebookId]/[paragraphId] "

Changed the doc as per code

What type of PR is it?

[Documentation]

Todos

What is the Jira issue?

ZEPPELIN-1562
How should this be tested?

Hit the url in below format for asynchrnous
http://[zeppelin-server]:[zeppelin-port]/api/notebook/job/[notebookId]/[paragraphId]

the same in synchronous
http://[zeppelin-server]:[zeppelin-port]/api/notebook/run/[notebookId]/[paragraphId]

Screenshots (if appropriate)

Questions:

Does the licenses files need update? NO
Is there breaking changes for older versions? NO
Does this needs documentation? NO

Author: Naveen Subramanian <nsubramanian-consultant@quixey.com>

Closes #1560 from snaveenp/ZEPPELIN-1562-api-doc and squashes the following commits:

169d9b6 [Naveen Subramanian] updated api doc